refactor(supervisor): split handlers by event family
Decompose the 1618-line supervisor/handlers/mod.rs into per-event-family

submodules to make the dispatcher path easier to audit.

New layout under src/supervisor/handlers/:

  - mod.rs    Dispatchers only: handle_dbus_event (D-Bus events) and

              handle_ipc_command (IPC commands), plus module declarations.

  - nm.rs     NetworkManager handlers: initial_nm_sync, poll_nm_state,

              force_state_resync, refresh_connections.

  - health.rs run_health_check.

  - user.rs   User-initiated commands: handle_connect, handle_disconnect,

              handle_restart, toggle_auto_reconnect, toggle_kill_switch,

              toggle_autostart, toggle_debug_logging, open_log_file,

              reload_configuration.

  - system.rs Lifecycle: handle_quit, graceful_shutdown, and the free

              resolve_restart_path helper (SHROUD-VULN-036 fail-closed).

All methods retain their original signatures and pub(crate)/pub(super)

visibility. Two methods (reload_configuration, resolve_restart_path) were

promoted from private to pub(super) so the dispatcher in mod.rs can still

reach them across the new module boundaries — call sites are unchanged.

Every SPDX header is preserved. Zero behavior change. cargo build /

clippy --all-targets --all-features -- -D warnings / test --all-features

all pass: 259+778+35+15+12 (baseline).
